Implemented ability to Add a new Team
The page to register a new team is up and the codebase works for
submitting a team. I don't currently pull back any players or regions,
so for now those values are hard coded in to allow a user to at least
enter into the database. Next step: pulling back players to allow a
captain to be chosen, and then pulling back regions (regions will be
easier than players due to there being a lot less)